Create a new set of "hidden" watch faces.
These watch faces are defined after the "secondary" ones but are skipped in normal Mode operation.

Instead, the intention is to have another watch face jump directly there, typically this can be used by the clock_face to jump to the set_time_face when Alarm is long pressed. As is done here.